03_Unit_3/Spotify/requirements.txt
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.8
Flask-SQLAlchemy==2.4.4
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.16.0
MarkupSafe==1.1.1
numpy==1.19.1
pandas==1.1.1
patsy==0.5.1
plotly==4.9.0
plotly-express==0.4.1
psycopg2==2.8.5
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
requests==2.24.0
retrying==1.3.3
scikit-learn==0.23.2
scipy==1.5.2
simplejson==3.17.2
six==1.15.0
sklearn==0.0
spotipy==2.13.0
SQLAlchemy==1.3.19
statsmodels==0.11.1
threadpoolctl==2.1.0
urllib3==1.25.10
Werkzeug==1.0.1